+++ clang-tools-extra/clang-tidy/ClangTidyForceLinker.h
@@ -103,6 +103,11 @@ extern volatile int PortabilityModuleAnchorSource;
 static int LLVM_ATTRIBUTE_UNUSED PortabilityModuleAnchorDestination =
     PortabilityModuleAnchorSource;
 
+// This anchor is used to force the linker to link the QuantumModule.
+extern volatile int QuantumModuleAnchorSource;
+static int LLVM_ATTRIBUTE_UNUSED QuantumModuleAnchorDestination =
+    QuantumModuleAnchorSource;
+
 // This anchor is used to force the linker to link the ReadabilityModule.
 extern volatile int ReadabilityModuleAnchorSource;
 static int LLVM_ATTRIBUTE_UNUSED ReadabilityModuleAnchorDestination =